pyproject.toml
[tool.poetry]
name = "movie-api"
version = "0.1.0"
description = ""
authors = ["Heo-JunHyeok <junhyeok_1996@naver.com>"]
readme = "README.md"
packages = [{include = "movie_api"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.6"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
